fix: type

+++ client/nodes.go
@@ -33,21 +33,21 @@ type Node struct {
 		SwapUsed           int     `json:"swap_used"`
 	} `json:"systemStats"`
 	InterestingStats struct {
-		CmdGet                       int `json:"cmd_get"`
-		CouchDocsActualDiskSize      int `json:"couch_docs_actual_disk_size"`
-		CouchDocsDataSize            int `json:"couch_docs_data_size"`
-		CouchSpatialDataSize         int `json:"couch_spatial_data_size"`
-		CouchSpatialDiskSize         int `json:"couch_spatial_disk_size"`
-		CouchViewsActualDiskSize     int `json:"couch_views_actual_disk_size"`
-		CouchViewsDataSize           int `json:"couch_views_data_size"`
-		CurrItems                    int `json:"curr_items"`
-		CurrItemsTot                 int `json:"curr_items_tot"`
-		EpBgFetched                  int `json:"ep_bg_fetched"`
-		GetHits                      int `json:"get_hits"`
-		MemUsed                      int `json:"mem_used"`
-		Ops                          int `json:"ops"`
-		VbReplicaCurrItems           int `json:"vb_replica_curr_items"`
-		VbActiveNumNonResidentNumber int `json:"vb_active_num_non_residentNumber"` // couchbase 5.1.1
+		CmdGet                       float64 `json:"cmd_get"`
+		CouchDocsActualDiskSize      int     `json:"couch_docs_actual_disk_size"`
+		CouchDocsDataSize            int     `json:"couch_docs_data_size"`
+		CouchSpatialDataSize         int     `json:"couch_spatial_data_size"`
+		CouchSpatialDiskSize         int     `json:"couch_spatial_disk_size"`
+		CouchViewsActualDiskSize     int     `json:"couch_views_actual_disk_size"`
+		CouchViewsDataSize           int     `json:"couch_views_data_size"`
+		CurrItems                    int     `json:"curr_items"`
+		CurrItemsTot                 int     `json:"curr_items_tot"`
+		EpBgFetched                  int     `json:"ep_bg_fetched"`
+		GetHits                      int     `json:"get_hits"`
+		MemUsed                      int     `json:"mem_used"`
+		Ops                          int     `json:"ops"`
+		VbReplicaCurrItems           int     `json:"vb_replica_curr_items"`
+		VbActiveNumNonResidentNumber int     `json:"vb_active_num_non_residentNumber"` // couchbase 5.1.1
 	} `json:"interestingStats"`
 	Uptime               string   `json:"uptime"`
 	McdMemoryReserved    int      `json:"mcdMemoryReserved"`
